1. Engine_Enum 파일 추가 2. typedef struct 에서 tag없는문제 모두다 해결 3. Engine_Defines에 필요한 헤더 추가 4. 유니티빌드 추가 5. 개인폴때 쓰던  rgba컬러 매크로 추가 6. Light_Desc 구조체 변경 - 생성시 index값을 리턴하게 만들음 -> Tick 혹은 Render, Light 죽을때 해당 오브젝트가 갖고있는 빛도 없애는 등등의 요소로 많이 사용되어서 넣음 able 여부 판단으로 Render 돌리도록함

+++ Engine/Private/Mesh.cpp
@@ -192,25 +192,25 @@ HRESULT CMesh::Ready_Vertices_Anim(const aiMesh * pAIMesh, const vector<class CB
 			/* pAIBone->mWeights[j].mVertexId : 이 뼈가 영향을 주는 j번째 정점의 인덱스 */
 			if(0.0f == pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.x)
 			{
-				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.x = i;
+				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.x = (uint32_t)i;
 				pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.x = pAIBone->mWeights[j].mWeight;
 			}
 
 			else if (0.0f == pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.y)
 			{
-				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.y = i;
+				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.y = (uint32_t)i;
 				pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.y = pAIBone->mWeights[j].mWeight;
 			}
 
 			else if (0.0f == pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.z)
 			{
-				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.z = i;
+				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.z = (uint32_t)i;
 				pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.z = pAIBone->mWeights[j].mWeight;
 			}
 
 			else if (0.0f == pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.w)
 			{
-				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.w = i;
+				pVertices[pAIBone->mWeights[j].mVertexId].vBlendIndices.w = (uint32_t)i;
 				pVertices[pAIBone->mWeights[j].mVertexId].vBlendWeights.w = pAIBone->mWeights[j].mWeight;
 			}
 		}
